// src/api/types.ts
export type MapType = 'boundary' | 'query' | 'table' | 'tileset' | 'raster';
export type RequestType =
  | 'Map data'
  | 'Map instantiation'
  | 'Public map'
  | 'Tile stats'
  | 'SQL'
  | 'Basemap style';
export type ScaleType =
  | 'linear'
  | 'ordinal'
  | 'log'
  | 'point'
  | 'quantile'
  | 'quantize'
  | 'sqrt'
  | 'custom'
  | 'identity';

export type APIErrorContext = {
  requestType: RequestType;
  mapId?: string;
  connection?: string;
  source?: string;
  type?: MapType;
};
// ...
export type Basemap = MapLibreBasemap | GoogleBasemap;